Rename the rake files
Their function has evolved over time so rename them to be more
representative of what they currently do (and move them into a separate
subdir)
