+++ rake_helpers/combine_sources.rb
